feat: Add draft of html
The base html should be done but will likely need to be tweaked as the
CSS is worked on. I put containers where I thought they would
be needed for flexbox. But the containers in particular are an area that
is subject to change.

Initial thought for the landing page: when people sign up, they get sent
short, daily emails with one space image, along with a brief explanation
of what the image is.