feat: Add CSS for hero section
Hero section should be done. Discovered something interesting about
buttons: it is not straightforward to add a link to them! I wanted to
add a link to the "Sign up" button so that it would theoretically do
something. There are options to use js or put it in a form tag. But my
conclusion was that I should not mess with that just yet. I'll see what
other students do with it in their examples and might ask a general
question about it in the discord.