fix: Change info image size to squares
Height was adjusted so that all the pictures in the info section are the
same size. The first photo, of the purple nebula, looks a little weird
now that it's been stretched. But I like the overall effect more. Now
the star cluster image isn't sticking out, taller than the other three.